+++ R/cTMed-delta-indirect-central.R
@@ -205,6 +205,10 @@ DeltaIndirectCentral <- function(phi,
                                  delta_t,
                                  ncores = NULL,
                                  tol = 0.01) {
+  idx <- rownames(phi)
+  stopifnot(
+    idx == colnames(phi)
+  )
   total <- FALSE
   args <- list(
     phi = phi,