+++ R/cTMed-mc-indirect-central.R
@@ -166,6 +166,10 @@ MCIndirectCentral <- function(phi,
                               ncores = NULL,
                               seed = NULL,
                               tol = 0.01) {
+  idx <- rownames(phi)
+  stopifnot(
+    idx == colnames(phi)
+  )
   total <- FALSE
   args <- list(
     phi = phi,